Add nuspec file for nuget package for GoIO SDK
This change adds a .nuspec file which creates a .nupkg which can be
uploaded to nuget.org.  A .NET developer can use the nuget package to
easily add references to the GoIO SDK.

The .targets file is used to update a Visual Studio project file to
automatically copy the native GoIO dlls so they are available at
runtime.

The readme contains notes and instructions for updating the nuget
package.

The nuget package is available online at
https://www.nuget.org/packages/GoIO.SDK/1.0.6

The version number is 1.0.6 because it took me six tries before I was
able to get the package to work correctly.